refactor: remove 'main-content.tsx'
